requirements.txt
openai
azure-cognitiveservices-speech
praat-parselmouth
